requirements.txt
blinker==1.4
captcha==0.4
certifi==2021.10.8
click==8.1.2
colorama==0.4.4
Flask==2.0.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-sqlalchemy==0.28.0
packaging==21.3
Pillow==9.1.0
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.8
six==1.16.0
SQLAlchemy==1.4.35
Werkzeug==2.1.1
wincertstore==0.2

openpyxl~=3.0.10
xhtml2pdf~=0.2.7